Fix the two CI guards the Infinity port left red
Both were red only in CI, because scripts/precheck.ps1 runs neither. That gap
is the same one the --isolate gas budgets fell through last week: a local run
that is greener than CI teaches you to trust it.

forge lint: the permission bitmap shifted a bare `1`, whose type comes from the
surrounding `uint16(...)` cast. The arithmetic was already right, and right for
a reason a reader has to reconstruct -- so `incorrect-shift` fired with no
recorded bound. Widened the operand to `uint16(1)` and wrote down what bounds
it: Infinity's offsets are uint8 constants topping out at 13, a uint16 has bits
0-15. The baseline is untouched, which is the point -- the finding is gone
rather than accepted.

checkV4RouterTuple: checks 1-4 were built around
`IV4Router.ExactInputSingleParams` being the production tuple. It is not, since
the port. Retargeted them at Infinity's `CLSwapExactInputSingleParams` and
demoted the Uniswap shape to what it now is -- the shape of the mistake, sent
at Infinity's router by test_forkInfinity_theUniswapShapedTupleIsNotInterchangeable.

A demotion, not a retirement. A trap has to be a faithful copy of the thing it
imitates or it proves nothing about the error anyone would actually make, so
UNISWAP_LAYOUT stays pinned to lib/, and a new check 2b pins the spike's
hand-rolled `UniShapedParams` to UNISWAP_LAYOUT. That closes the loop the file
was written about: an unrelated copy. Check 1 stopped demanding an import,
because Infinity's type cannot be imported without a remapping that moves every
production metadata hash; check 6 is what makes the restatement safe.

Note that the bitmap change moves the hook implementation's creation code, so
CREATE2 addresses shift. Nothing is deployed on BSC yet.

387 tests pass under --isolate; both guards green without relaxing either.

+++ src/ToshLaunchpadHook.sol
@@ -2222,12 +2222,26 @@ contract ToshLaunchpadHook is ICLHooks, ILockCallback, ReentrancyGuard {
     ///         Declared `public` rather than `external` so `_key()` can read it
     ///         without a self-call: the key has to repeat the bitmap, and the
     ///         two must come from one expression or they can drift apart.
+    ///
+    ///         WHAT BOUNDS THE SHIFTS, since `forge lint` asks and the answer is
+    ///         not visible from here. Each `HOOKS_*_OFFSET` is a `uint8` constant
+    ///         in Infinity's `ICLHooks`, and the largest of the fourteen is 13
+    ///         (`AFTER_REMOVE_LIQUIDIY_RETURNS_DELTA`). A `uint16` has bits 0-15,
+    ///         so every shift Infinity defines fits with two bits to spare and no
+    ///         combination of them can truncate.
+    ///
+    ///         The shifted operand is written `uint16(1)` rather than a bare `1`
+    ///         for that reason. A literal takes its type from context, which here
+    ///         is an outer `uint16(...)` cast — so the arithmetic was already
+    ///         correct, and correct for a reason a reader has to work out. Naming
+    ///         the width at the shift makes the bound local: if Infinity ever
+    ///         defines an offset above 15, this stops compiling silently-correct
+    ///         code and starts truncating visibly at a named type.
     function getHooksRegistrationBitmap() public pure override returns (uint16) {
-        return uint16(
-            (1 << HOOKS_BEFORE_INITIALIZE_OFFSET) | (1 << HOOKS_BEFORE_ADD_LIQUIDITY_OFFSET)
-                | (1 << HOOKS_BEFORE_SWAP_OFFSET) | (1 << HOOKS_AFTER_SWAP_OFFSET)
-                | (1 << HOOKS_BEFORE_SWAP_RETURNS_DELTA_OFFSET) | (1 << HOOKS_AFTER_SWAP_RETURNS_DELTA_OFFSET)
-        );
+        return (uint16(1) << HOOKS_BEFORE_INITIALIZE_OFFSET) | (uint16(1) << HOOKS_BEFORE_ADD_LIQUIDITY_OFFSET)
+            | (uint16(1) << HOOKS_BEFORE_SWAP_OFFSET) | (uint16(1) << HOOKS_AFTER_SWAP_OFFSET)
+            | (uint16(1) << HOOKS_BEFORE_SWAP_RETURNS_DELTA_OFFSET)
+            | (uint16(1) << HOOKS_AFTER_SWAP_RETURNS_DELTA_OFFSET);
     }
 
     /// @dev Settle the full-range genesis position: `lpNative` native ETH plus
